# src/NileTechno.API/openapi/niletechno-openapi.yaml
openapi: 3.0.3
info:
  title: NileTechno API
  description: |
    Frontend API Reference for NileTechno E-Commerce API.
    JSON fields use camelCase. Public product/catalog data only includes active ERP products/groups.
  version: 1.0.0
servers:
  - url: https://my-back-end-jayl.onrender.com
    description: Production
  - url: http://localhost:5080
    description: Local Development
...
security:
  - {}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token. Expires after 60 minutes.
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success: { type: boolean }
...
    CartProduct:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        title: { type: string }
        price: { type: number, format: double }
        image: { type: string }
        stock: { type: number, format: double }
        groupId: { type: string }
        category: { type: string }
...
    WishlistItem:
      type: object
      properties:
        id: { type: string }
        productId: { type: string }
        product: { $ref: '#/components/schemas/CartProduct' }
    WishlistRequest:
      type: object
      required: [productId]
      properties:
        userId: { type: string, description: يُتجاهل لغير الأدمن — الهوية من التوكن }
        productId: { type: string }
    WishlistSyncRequest:
      type: object
      required: [productIds]
      properties:
        userId: { type: string, description: يُتجاهل لغير الأدمن — الهوية من التوكن }
        productIds:
          type: array
          items: { type: string }
...
paths:
...
  /api/wishlist:
    get:
      tags: [Wishlist]
      summary: Get wishlist
      security: [{ bearerAuth: [] }]
      description: أي صنف Status=0 (أو مجموعته موقوفة) لا يظهر في القائمة — هو غير موجود من وجهة نظر العميل.
      parameters:
        - { name: userId, in: query, required: false, schema: { type: string }, description: للأدمن فقط — لغير الأدمن تتجاهل السيرفر القيمة وتستخدم هوية التوكن }
      responses:
        '200':
          description: Wishlist
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/WishlistItem' }

    post:
      tags: [Wishlist]
      summary: Add item to wishlist
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WishlistRequest' }
      responses:
        '200':
          description: Added
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }

    delete:
      tags: [Wishlist]
      summary: Remove item from wishlist
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: userId, in: query, required: false, schema: { type: string }, description: للأدمن فقط — لغير الأدمن تتجاهل السيرفر القيمة وتستخدم هوية التوكن }
        - { name: productId, in: query, required: true, schema: { type: string } }
      responses:
        '200':
          description: Removed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
  /api/wishlist/sync:
    post:
      tags: [Wishlist]
      summary: Replace wishlist contents
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WishlistSyncRequest' }
      responses:
        '200':
          description: Synchronized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
...